Mudança para float

+++ src/main/java/com/github/fabriciofx/ta/banco/Conta.java
@@ -6,5 +6,5 @@ public interface Conta {
 	
 	void saque(double valor);
 	
-	void deposito(double valor);
+	void deposito(float valor);
 }
